+++ src/components/ui/button.tsx
@@ -1,15 +1,10 @@
 import * as React from 'react'
 import { Slot } from '@radix-ui/react-slot'
-import { type VariantProps } from 'class-variance-authority'
 
 import { cn } from '../../lib/utils'
 import { buttonVariants } from './button-variants'
 
-export interface ButtonProps
-  extends React.ButtonHTMLAttributes<HTMLButtonElement>,
-    VariantProps<typeof buttonVariants> {
-  asChild?: boolean
-}
+import type { ButtonProps } from '../../types'
 
 const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   ({ className, variant, size, asChild = false, ...props }, ref) => {
