+++ src/routes/Dashboard.tsx
@@ -8,7 +8,7 @@ import {
 } from '../components/ui/card'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '../components/ui/tabs'
 import { useBugStore } from '../store'
-import { Bug } from '../types/bug'
+import { Bug } from '../types'
 import { Badge } from '../components/ui/badge'
 import { Progress } from '../components/ui/progress'
 import { Button } from '../components/ui/button'
