--- src/types/bug.ts
@@ -1,11 +1,0 @@
-export interface Bug {
-  id: string
-  title: string
-  description: string
-  bounty: number
-  active: boolean
-  priority?: 'high' | 'medium' | 'low'
-  createdAt?: string
-  resolvedAt?: string
-  assignee?: string
-}
