--- src/types/user.ts
@@ -1,12 +1,0 @@
-import type { Bug } from './bug'
-
-export interface User {
-  id: number
-  name: string
-  avatar: string
-  bugs: Bug[] | number[]
-  bounty: number
-  score?: number
-  bugsSquashed?: string[]
-  survivor?: boolean
-}
